bugfix(schematics): match on "from" instead of "import" to avoid breakage of multi-line import
Currently, the matching for generating module imports is done using a
regex for "import {" statements. When existing code contains one or more
multi-line imports, as is often found in corrections from Prettier, the
new line breaks TypeScript syntax by being inserted in the middle of
the existing multi-line imports. This corrects that issue by matching on
"from" instead. Also, there are changes to the algorithm to insert
generated import stamtement via Array#splice instead of using concat on
multiple arrays.

Closes #204