feat(application): enable strict mode for typescript
Strict mode is a good default for new projects, because it prevents
accidental omission of type safety information. It's much easier to
maintain strict mode compatibility from the start than it is to
retrofit a mature project to support strict mode.

That said, it's useful to default strictPropertyInitialization to
false for Nest projects, since the class-validator DTO pattern
described in the documentation is not compatible with strict property
initialization.